rename to add plan view
